Simplified and split constraints_library into constraint_approximation
