+++ moveit_planners/ompl/ompl_interface/CMakeLists.txt
@@ -9,11 +9,12 @@ add_library(${MOVEIT_LIB_NAME}
   src/parameterization/joint_space/joint_model_state_space_factory.cpp
   src/parameterization/work_space/pose_model_state_space.cpp
   src/parameterization/work_space/pose_model_state_space_factory.cpp
-  src/detail/constraints_library.cpp
   src/detail/threadsafe_state_storage.cpp
   src/detail/state_validity_checker.cpp
   src/detail/projection_evaluators.cpp
   src/detail/goal_union.cpp
+  src/detail/constraints_library.cpp
+  src/detail/constraint_approximation.cpp
   src/detail/constrained_sampler.cpp
   src/detail/constrained_valid_state_sampler.cpp
   src/detail/constrained_goal_sampler.cpp
